+++ src/web/comment/store/commentStore.ts
@@ -10,8 +10,8 @@ import { apiSyncer } from "@/web/comment/store/apiSyncerMiddleware";
 import { emitter } from "@/web/common/event";
 import { storageWithDates } from "@/web/common/store/utils";
 import { StoreTopic } from "@/web/topic/store/store";
-import { setSelected } from "@/web/view/currentViewStore/store";
 import { toggleShowResolvedComments } from "@/web/view/miscTopicConfigStore";
+import { setSelected } from "@/web/view/selectedPartStore";
 
 export type StoreComment = Omit<Comment, "topicId">;
