+++ src/web/common/components/ContextMenu/ViewDetailsMenuItem.tsx
@@ -1,7 +1,7 @@
 import { ContextMenuItem } from "@/web/common/components/ContextMenu/CloseOnClickMenuItem";
 import { emitter } from "@/web/common/event";
 import { GraphPart } from "@/web/topic/utils/graph";
-import { setSelected } from "@/web/view/currentViewStore/store";
+import { setSelected } from "@/web/view/selectedPartStore";
 
 export const ViewDetailsMenuItem = ({ graphPart }: { graphPart: GraphPart }) => {
   return (
