+++ src/web/topic/components/Indicator/CommentIndicator.tsx
@@ -6,8 +6,8 @@ import { useCommentCount } from "@/web/comment/store/commentStore";
 import { emitter } from "@/web/common/event";
 import { Indicator } from "@/web/topic/components/Indicator/Indicator";
 import { GraphPartType } from "@/web/topic/utils/graph";
-import { setSelected } from "@/web/view/currentViewStore/store";
 import { useShowResolvedComments } from "@/web/view/miscTopicConfigStore";
+import { setSelected } from "@/web/view/selectedPartStore";
 
 interface Props {
   graphPartId: string;
